make fetch_api a python module
