+++ fetch_demos/CMakeLists.txt
@@ -2,3 +2,4 @@ cmake_minimum_required(VERSION 3.7.2)
 project(fetch_demos)
 find_package(catkin REQUIRED)
 catkin_metapackage()
+
